Move tracing basics to contrib folder
